technical_case/requirements.txt
Django>=4.2.2
djangorestframework>=3.10.3
djangorestframework-gis>=0.14
psycopg2>=2.9.6